Add support for configurable lookup paths
In a large project you often have multiple places where js files may
be located. In a project I'm working on right now, we have the following
paths setup in our webpack.config.js:

  __dirname + '/app/assets/javascripts',
  'vendor/assets/bower_components',
  'node_modules',
  'spec/javascripts'

I'm adding support for a .importjs config file. This file should be in
YAML format, and the only configuration supported right now is
`lookup_paths`, an array of paths to look for js files in on importing.

The .importjs file should be placed in the root of your project.