Resolve multiple matches in an inputlist
Instead of just picking the first file if multiple files resolve to the
variable name, we now show a list of all matches and lets the user pick
one for herself.

After doing some digging, I decided that using `inputlist` was the
simplest method to get this done. There are other ways of bringing up a
menu, but most of them require bringing in a external dependency or
coding it yourself.