Use Ruby 1.8 syntax for method chaining in `importer.rb`
Support for newlines after chained method calls was added to Ruby in
version 1.9.[1] On my system, the default installation of Vim is using
an older version of Ruby. To support pre-1.9 versions of Ruby, we can
change the placement of the dot in the method calls.

Recent Ruby style guidelines prohibit this syntax,[2] but in this case
it allows more users to easily try out this Vim plugin.

[1]: http://ruby.about.com/od/newinruby191/qt/dotsyntaxchange.htm
[2]: https://github.com/bbatsov/ruby-style-guide#consistent-multi-line-chains