Add command to import all undefined variables
A couple of times, I've been in a situation where I wanted to import all
undefined variables in a file. Lately, I've been using a tool called
jshint to find issues with my js code. Jshint can find unused variables.
So I made a small integration with the jshint command line tool, where
I send the current buffer to it, parse the response for unused variables
and then import those variables.

Right now, the tool is hard-coded to run `jshint` from the command-line,
assuming a global install. This is not ideal, and I intend to follow up
with a configurable path to run it.