Add declaration_keyword configuration option
If you are using ES6 (ES 2015), you have access to `let` and `const` in
addition to `var` as ways to declare variables. `const` seems like an
appropriate declaration keyword to use when requiring modules, so it
would be nice if this plugin supports it.

This commit adds support for a `declaration_keyword` option that can be
set to a string that will be used instead of `var` when declaring the
variables that are assigned to the required modules.