Teach finding and uniqing imports about const and let
In ES6 (ES2015), you can declare variables using `const` and `let`
keywords in addition to the trusty `var`. This commit teaches import-js
about these keywords in methods that find the current set of imports and
in determining uniqueness of the found imports. This allows people to
use these new keywords with this plugin.