Auto-break long lines
It would be great if I could specify a character limit (defaulted to 80)
that import-js would try to stay within when importing new dependencies.

So instead of import-js generating this:

  const MyModule = require('path/to/my_module');
  const MyModuleThatHasALongName = require('path/to/my_module_that_has_a_long_name');

it would generate this:

  const MyModule = require('path/to/my_module');
  const MyModuleThatHasALongName =
    require('path/to/my_module_that_has_a_long_name');

This commit makes this happen by adding a `text_width` configuration
option that defaults to 80. I borrowed this name from Vim's `textwidth`
option.

This closes #12. It would be nice to follow up at some point with a way
to configure (or even auto-detect) the indentation style. For now, I
just went with what I use, which is 2 spaces.