Fix inserting multiline imports into buffer again
Before this change, they would end up as

  var SomeVariable =~@  require('some_variable');

instead of

  var SomeVariable =
    require('some_variable');

`buffer.append` wasn't dealing with "\n" newlines the way I expected.
This wasn't caught in specs because we use a mock buffer and not a real
one. An issue for this has been created at:

  https://github.com/trotzig/import-js/issues/13

This was originally fixed by 5f91595a, but I accidentally clobbered it
when I tried to make this method more readable. I added a comment to
clarify why this is written the way it is.

I decided to go back to the reverse_each to prevent having to bookkeep
the line number with the additional lines.