Make mock buffer behave more like a real buffer
If you try to append a string with \n newline characters in it, a real
vim buffer will output "^@" in place of the newline characters. The
mock buffer that we've been using didn't do that which led to a bug
fixed in https://github.com/trotzig/import-js/commit/5f91595aa4. Making
the mock behave more like a real buffer will prevent us from making this
mistake again.

Fixes #13