Add YARD to MockVimBuffer, alias length to count
The YARD makes this object easier to understand. I pulled most of the
descriptions from

  http://vimdoc.sourceforge.net/htmldoc/if_ruby.html#ruby-buffer

This class is getting close to being fully featured and might make a
good candidate for releasing as its own gem in addition to
MockVimWindow.

Alternatively, it might be nicer to switch to something vimrunner for
more accurate testing.

  https://github.com/AndrewRadev/vimrunner

Here is a tutorial that explains how to set that up with Travis CI:

  http://mudge.name/2012/04/18/testing-vim-plugins-on-travis-ci-with-rspec-and-vimrunner.html