Replace text_width config with textwidth detection
In 75557f4e6f we added a feature that auto-breaks long lines based on
the `text_width` option in the .importjs configuration file. This is
less than ideal because it requires extra effort to get it configured.
Instead, we can just use Vim's textwidth setting.

For this work, I swiped the `exists?` and `get_number` methods from
@Wincent "international man of mystery" Colaiuta's Command-T[0]. There
may be a better place for this, but for now I figured this was good
enough.

[0]: https://github.com/wincent/command-t/blob/13760a7/ruby/command-t/vim.rb#L15