Extract configuration into its own class
Loading the configuration was intermingled with the rest of the code in
Importer.rb. I wanted that to become a bit less cluttered, and at the
same time allow for easier configuration refactorings. So I extracted
the config parts into a new class, "ImportJS::Configuration".

I also moved all the config related methods added to Importer.rb in
590e7d5ae28b by @lencioni to the new Configuration class.