Show error if YAML configuration file is found
This will hopefully help others who have already installed this plugin
and have created a YAML file called ".importjs".

I decided to add it as early as possible in the code and keep it
separated from the rest of the config. I didn't add any tests either,
which was a conscious move. Hopefully this won't be long-lived.