Scan through all .js* files on import
In order to get better control over what is imported, I've wanted to
simplify the Dir.glob call and do more post-processing to select exactly
the files to import. At first, I tried to find a glob pattern that would
match both files directly and node.js conventional modules (e.g.
modules/FooModule/index.js). The closest I could do was this

  **/foo*module**.js

This pattern would match e.g. bar/foo_module.js.jsx (regular file), but
not bar/foo_module/index.js (node.js module).

Instead of trying to find the perfect glob pattern, I reverted to
instead scan through all **/*.js* files on the first pass. Each file
matched by that pattern will then go through a filter to look for
exactly the string we're looking for. This is of course not great for
performance, but after playing around with it for a while I think it's a
decent trade-off.

A logical next step is probably to cache the initial result from
Dir.glob. I'm hesitant to introducing a cache at this moment though as I
think it might make people loose faith in the tool (stale caches are the
worst -- it's up there with patriarchy, child soldiers and over-ripe
mango).

The spec got a little bit better now that we do most of the filtering in
ruby. Before, we had mocked out most of the core of importing. Now, we
only give it the full list of files to scan.

While in the spec file, I did some clean-up from what Rubocop told me
was bad.

Inside the formatted_to_regex method, I decided to get rid of a
conversion from :: to /. I think this is a remainder from a copy-paste,
and is not something that we need.