Add caching layer
This commit is an attempt to speed up importing the second, third,
fourth, and so forth, files. A cache is populated for each path defined
in `lookup_paths`, and if nothing has changed (looks for modified
folders) since last fetch, the cached result is reused.

Through manual testing, I think this takes down the time spent on the
second import by roughly half the time. The reason it isn't super fast
is that we still perform a check to see if any folder (recursively) has
changed since last check. This is important to avoid stale caches, which
is something we don't want.