Match directories that have `main` defined in `package.json`
The package.json file of a node module can (and often does) contain a
`main` property pointing to the main module of a package. For instance,
the package.json of the `json-sass` package points to "lib/jsonSass.js".
To import json-sass, all you have to do is `require('json-sass')`.
Webpack and other node-compatible tools will figure out what file to
really look for based on the information in the package.json file.

There's one side-effect of how I implemented this behavior that I hope
to address soon: when multiple files resolve to a variable and one of
them is derived from the package.json file, you will see something like
this:

  [import-js]: Pick file to import for `Foo`:
  1. Foo/package.json
  2. bar/foo.js

It would be better if #1 said "Foo", or perhaps the entire path to the
Foo folder, "app/assets/components/Foo". I plan on tackling that later.

Fixes #27