Allow aliases to contain destructured objects
This is useful if you are using a library that exports a single object
but that objects has properties (objects, methods) that we want to
single out in our import. Here's an example with Underscore.js:

  const { memoize } = require('underscore')

  memoize(() => { doSomethingExpensive() })

The feature requests mentions combining multiple destructured objects
into one, but that turned out to be a little bit too involved for the
time I have available. A future enhancement could make sure that
multiple destructured objects are combined into one:

  const { memoize, debounce } = require('underscore');

Fixes #38