Simplify Importer class
There was some old tech debt in the complexity involved in knowing
whether or not an import was added. A while back, this was used to
determine if the cursor should move a line after importing or not. This
changed some time ago to instead count the lines in the buffer before
and after. But the code to signal added or not was still around. And it
was only used in one place: to display a message after
:ImportJSImportAll was run. That message was redundant anyway since we
show individual messages per variable imported, so I removed it.