Fix bug with organizing imports
If you had a buffer that looked something like this

  var a = require('a');
  var b = a.b;
  var c = require('c');

the buffer would be changed to this on successful import:

  var a = require('a');
  var c = require('c');

  var b = a.b;
  var c = require('c');

Notice the duplicate `c` import. The bug was probably introduced in
1cc2157 when I added support for multiline imports. I change the code in
find_current_imports to stop scanning for imports if it finds something
that is not an import. The result is that import-js will become better
at breaking off the list of imports from the rest of top-level
initialization, but if you start mixin imports and regular assignment at
the top you might be surprised with the result. For instance, the
example shown above will result in this output (let's pretend that
you're importing `d` here):

  var a = require('a');
  var d = require('d');

  var b = a.b;
  var c = require('c');

This output is partly a result of what was easy to implement, and a
future enhancement might change the implementation to generate this
output instead (again, let's pretend we're importing `d`):

  var a = require('a');
  var c = require('c');
  var d = require('d');

  var b = a.b;