Dedupe modules when lookup_paths overlap
At Brigade, we have two lookup_paths that overlap:

  app/assets
  app/assets/javascripts

When you try to import a module in `app/assets/javascripts`, you get two
options:

  [import-js] Pick js module to import for 'warn': (0.24s)
  1: javascripts/lib/warn
  2: lib/warn

The overlapping lookup paths are for many reasons not ideal, but
changing this structure isn't a straightforward change. So I've decided
to do something about it in import-js instead.

Assuming that we always want the shorter import path (`lib/warn` in the
example above), we first sort modules by import path length, then dedupe
modules. This will remove the module(s) with the longer path(s). We then
sort again, to get consistent alphabetic sorting in the selection list.