Add :ImportJSGoTo to quickly open a file
This is something that I've wanted to try for a while. Since Import-JS
is already pretty good at resolving a variable name to a file, we might
as well add an option to quickly go to (open) that file.

This feature is still experimental. I haven't added any tests, and if
you end up having to resolve multiple modules in a selection list, the
copy will talk about "importing" that variable. If this feature plays
out well, I might polish it.

Now, some might say "just use `gf`". That's certainly an option, it's
just that I find it tricky to configure correctly. And it wasn't hard to
implement in import-js (I know that's not a great reason for adding
features in general though...).