Extract VIM.message calls to private message()
We want to truncate all messages being sent to Vim to the width of Vim
to prevent people having to press enter to continue. To make this
easier, I think it will be useful to have our own message function that
we call that delegates to the VIM.message function. This way we can do
the truncation in one place instead of many.