Truncate messages to prevent having to confirm
Sometimes, when you're importing a module with a long name/path to, Vim
isn't wide enough to display the full message. In those cases, you have
to hit Enter in order to continue doing what you were doing in the file.

In this commit I fix this by looking at the width of Vim and
automatically truncating any messages being displayed.

I'm not entirely sure how Vim decides to require the user to press
Enter, but we might be able to avoid truncating by doing some manual
wrapping. However, since I know this will fix the issue and it is a
simple solution, I am going for this for now.

Fixes #48