DRY up messages
Now that we are funneling every message through the same helper method,
we can DRY up the beginnings of the messages and just add the
"[import-js]" tag in one place. While I was at it, I decided to drop the
":" from this since it seems unnecessary with the brackets and it saves
us 1 character.