Use WideMsg function to allow messages to be wider
After truncating messages and trying it out in my Vim, I noticed that it
didn't always solve the issue that we were trying to solve: avoiding the
press Enter prompt. After a bit of looking at `:help hit-enter`, I
discovered that the prompt takes into account the `showcmd` and `ruler`
options.

I found a function on the World Wide Web [0] that temporarily disables
these options to show a wide message. Once that was in place, I simply
needed to update the specs to match.

While I was manually testing this, I noticed that there needs to be an
empty space at the end of the line to avoid the prompt, so I shortened
the message by another character.

[0]: http://vim.wikia.com/wiki/How_to_print_full_screen_width_messages