Import directly via package.json
As your node_modules directory grows, it can become harder to prevent
the wrong modules from being imported. Npm will automatically flatten
sub-dependencies (foo depends on bar, baz depends on bar) into top-level
modules, and it isn't easy to spot that an import is using such an
indirect dependency. You might see this in your project

  const alpha = require('alpha');

even though "alpha" is not listed as a dependency in your packages.json
file.

To prevent this from happening, and to make import-js more closely tied
to package.json, import-js will now parse package.json for
"dependencies" and "peerDependencies" and import such modules if they
match the variable to import (given that they have a `main` module
defined inside their package.json).

A positive side-effect from this change is that performance is way
better now that we don't have to traverse the entire node_modules
folder.

It is recommended that you remove `node_modules` from the lookup path if
you have a package.json file to describe your dependencies. If you
don't, you won't get the benefits outlined in this commit (importing
will still work though).

Thanks to @lencioni for thinking about this and coming up with great
ideas that led to this commit.

Fixes #43