Move out VIM dependency from ImportJS::Importer
I want to make this tool available for non-VIM users as well. I plan to
do that by turning import-js into a Ruby gem that others can depend on.
Before that's useful however, I need to detach the direct VIM dependency
from the importer. This commit does that by moving all the VIM
communication into a ImportJS::VIMEditor class. This class will define
the API needed in order to integrate with another editor.