Add command-line "editor"
This is mostly a POC to demonstrate that import-js can be used in other
environments than only VIM. The CommandLineEditor will use an array of
strings to hold the file contents and operate on it. It can import one
specific variable or all undefined variables.

Invoke it like this:

  cat some-file.js | import-js Button

to import `Button` inside some-file.js

  cat some-file.js | import-js

to import all undefined variables inside some-file.js

I haven't figured out what to do with the end result, so right now I
print the entire file to stdout. The idea is that you should be able to
direct stdout to a file, but right now that doesn't work because logs
are also written to stdout. I'll investigate more how to do that, but at
least this will demonstrate how to write a custom editor integration.