Add basic Emacs support
This has a long way to go to become a full featured setup for import-js,
but it is functional.

Adds two pieces:
1. ImportJS::EmacsEditor class that uses the ImportJS APIs to resolve
modules and add `require` lines.
2. An elisp plugin to start up and talk to a Ruby process that runs the
ImportJS::EmacsEditor.

For now, this works a little differently from the Vim plugin. We don't
have direct access to the Ruby code in our buffers; instead, we
communicate with the process using stdin/stdout, and leave the file
handling to Ruby.

I think there's some additional optimizations to be made by further
breaking out the core import functionality from the original Vim plugin,
and also implementing the 'experimental' goto interface.