Replace `keep_file_extensions` bool with `strip_file_extensions` array
In order to get better control over what the resulting `require`
statement ends up as, we need more fine-grained control over what file
extensions are stripped out. I recently ran into a problem importing a
`.json` file because or config was set to strip out all file extensions.

The old configuration option was

  "keep_file_extensions": false

The new option is

  "strip_file_extensions": ['.jsx', '.js']

The dot (.) is required, and I choose to include it because that's what
webpack is doing [1]. (Thanks to @lencioni for linking me there.)

While I was in specs, I decided to rename a few .js.jsx extensions so
that we could make use of the default configuration for
`strip_file_extensions`.

Fixes #62

[1]: https://github.com/webpack/docs/wiki/configuration#resolveextensions