Add ability to import using `import from` statements
This is quickly becoming the default way to import modules in ES2015
projects:

  import foo from 'foo'

React-router [1], Airbnb [2] and others use it.

I decided to use the existing `declaration_keyword` to begin with. It
isn't necessarily a "keyword" per se, but I think the README can
explain it well enough so that it doesn't become an issue.

[1]: https://github.com/rackt/react-router
[2]: https://github.com/airbnb/javascript#modules