Fix de-duplicating modules that have "from" in them
In 50d3bf1433 we added support for ES2015 modules. Upon reviewing the
code, I spotted a potential problem for modules that have "from" in them
[0] because of how the regexes were being applied. This commit aims at
fixing this problem and adds a spec that will prevent a regression.

[0]: https://github.com/trotzig/import-js/commit/50d3bf14#commitcomment-14768482

I pulled the normalization out into its own method and broke the regexes
out into multiline regexes with some comments to make them more
readable. Although it is a bit of an implementation detail, it might not
be a bad idea to test the normalize_import method directly so that we
can have more thorough coverage here.