Pass around `ImportJS::ImportStatement`s instead of strings
This removes the need for parsing import strings in more than one place
which should make the code more robust.

To avoid rewriting import strings for already existing imports, I
decided to cache the original import string and reuse it in
`to_import_string`. This is something that we might consider removing
later. The result would be that all import statements are subject to
change as part of importing a variable.