Raise error when eslint/jshint parsing fails
We currently have a bug with the comment added for eslint parsing,
leading to a parse error when eslint or jshint is run. To make it more
clear when eslint/jshint fails to parse the file, I'm raising an
exception when such a scenario happens. This will show up as a red error
message in VIM.