Remove support for jshint
Since not many people have complained about the bug with import_all that
we've had for a while (syntax error introduced in an eslint comment), I
think it's safe to say that this feature is mostly unused. Supporting
both jshint and eslint has started to become a paing, so I'm making the
call to remove support for jshint. I also removed the configuration
option to customize the command, which will hopefully lead to less
complication down the road.