Prevent unnecessary import-rewriting
Before, when importing all variables we would rewrite import lines for
each variable. That was time-consuming, and led to some UI freeze if you
were importing a large number of variables. I've moved some code around
and renamed a few things so that all imports are resolved before finally
writing them to the editor.