Use --rule flag when running ESLint
Now that we only use ESLint to find unused variables, we can work with
the ESLint API more directly. This means that instead of modifying the
contents that ESLint looks at, we can use CLI flags. This feels a bit
cleaner to me.

We could probably go a step further and pass the path to the file
instead of passing it through stdin_data, but that would require
modifying the interface to Editor and the value isn't very clear to me
at the moment. Perhaps this would allow us to remove the dependency on
Open3 but since that is part of Ruby stdlib, I'm not sure that really
matters much.