Fail on all ESLint parsing errors
I am working on debugging #56 and noticed that in the simple project I
set up to work on this issue, I am seeing the following error when
running ESLint:

> Parsing error: Illegal export declaration

I considered adding this exact message to the list of error messages we
check for and fail on, but I decided that there are likely a bunch of
other parsing errors in the wild that we would still miss. It seems that
we'd want to fail on any parsing error, since this seems to be a
showstopper for ESLint.