Check for ESLint SyntaxError in stderr
While investigating #56, I managed to make my .eslintrc.js file
unreadable by ESLint. Instead of telling me that this was happening, I
ended up with a silent failure mode where ImportJS told me that there
are no variables to import. To make it more obvious what is happening, I
decided to add a check for this and fail loudly.

I considered just having it fail if stderr contained any output, but I'm
not entirely sure how ESLint uses stderr so I decided to err on the side
of caution for the time being.