Add no-undef to ESLint rules
The no-unused-vars rule was checking for variables that were defined but
not actually used. This was useful for the recently added functionality
that removes imports that are no longer used, but is not helpful for
finding imports that are needed. To make this work correctly on a
project that doesn't already have this rule enabled, I am adding the
no-undef rule to find these types of things. This will make the
auto-importing work better for more types of projects.

http://eslint.org/docs/rules/no-undef.html

Fixes #56.