Add ability to use `{filename}` in aliases
By using {filename} in an alias, you can use the name of the currently
edited file in the resulting import statement. @lencioni suggested
looking at vim-projectionist [1] for inspiration. Which I did, but I
decided to not use much of it. The main reason being that projecting is
a different usecase, where you have configuration that apply on a large
set of files. We want to apply a configuration on a single alias.

To keep it simple, I only added the filename (without file extension) to
begin with. I can imagine usecases where you need more than just that
though (e.g. a relative path within the project, or a relative path from
a `lookup_path`). We can extend this functionality when those usecases
pop up.

[Fixes #76]

[1]: https://github.com/tpope/vim-projectionist