Add support for selecting from a list in Sublime
This was trickier than I first thought, but hopefully it's worth the
complexity. When an import can't be resolved, we normally ask the user
to select something. When using the command line editor however, we
can't make this synchronous. To solve that, I've implemented a way to
rerun a command, but with a `--selections` argument passed in. The
original selections to choose from come from the metadata communicated
in the stderr pipe as json. I first tought about sending json back in
the call to import-js, but that turned out to be complicated, so I
decided to use a comma-separated list of <word>:<selected_index> tuples
instead.

I changed the signature of `ask_for_selections` so that the
word/variable you are about to import is included. This made rerunning
the command with --selections easier.