Add --goto option for CLI interface and use in Sublime
By passing in this option, a path to the js module will be printed to
stdout instead of the full file content. This string is then used to "go
to" (open) a file in Sublime.

If multiple modules map to a word, the regular selection list will
appear and you will have to select something before the file is opened.

While doing all of this, I found a bug in the goto method: if no js
module was resolved, then a nil pointer thing would happen.