Add fix_imports API to Emacs plugin
Allow users to use the 'fix_imports' function that automatically adds
unimported variables and removes unused imports. Fix a literal '\n' to
get file contents.

Add instructions on setup. The setup can be a little dicey and is
dependent on your eslint config, which I think is a bit dangerous, so
calling it 'optional'.

Remove references to 'experimental' for Emacs (its not really
experimental anymore =D).