Revert "Warn on plugin load if executable missing or wrong version"
This reverts commit 0f7d1314eab07f2f7948d2897bde2139673ac153.

It turns out that Package control bundles the plugin in a single file
called ImportJS.sublime-package. So checking for files won't really work
unless we unpack the plugin. Which we don't want.

I'm going to mention having to `gem update import_js` any time a new
version is installed instead.