Improve Travis CI build matrix
YAML is confusing, but we don't actually need quote marks here so I
removed them. I also changed 2.1.1 and 2.2.1 to the more general 2.1 and
2.2, which will use the latest patch version of each minor version. This
seems preferable since most people will probably be tracking at least
the patch versions. I also noticed that 2.0.0 was not covered, so I
added that to fill the gap.