Rename ImportStatement#inject_{,destructured_}variable
I am working on #83 and I think I see a path to making this work that
will require separating variables into default_variable and
destructured_variables. To begin that work, I am going to start renaming
some things to improve the clarity of what they are supposed to be
doing.