Separate default and destructured variables in ImportStatement
I'm working on #83 where we will need to be able to generate import
statements that contain both default and destructured variables. This
commit brings us a step closer to that goal by separating how
ImportStatement represents this information into two separate variables.
I think that even if we weren't working on this issue, this is cleaner
and a more accurate way to represent this information which leads to
better grokability of the code.