Merge found imports
In my previous commit, I improved the stability of sorting of the
imports so that they wouldn't shift on subsequent runs. After giving it
some thought, I realized that ImportStatements should be pretty stable
even if they are returning multiline strings so I dug into it some more.

It turned out that when finding old imports, we weren't making any
effort to merge two imports that were importing the same module into the
same ImportStatement. Doing so will allow us to be more flexible with
how we choose to output the import block, which I see as a win. For
instance, we can now sort the destructuring right next to the original
importing if we want. This should also make it easier to switch our
default variable + destructuring variables in CommonJS syntax to not
require the module twice if we want.

One potential downside here is that if you are importing the same module
twice but with different variable names, only one will win and the other
will be discarded. e.g.:

  var Foo = require('foo');
  var Bar = require('foo');

will likely become something like:

  var Bar = require('foo');

This may cause some problems but it is not a good practice in the first
place, so I actually see this as a feature.