Extract ImportStatement#line_too_long?
I am working on #106 where I will need to use different rules for
wrapping imports with default variables vs imports with destructuring.
To make this a little more manageable, I am extracting a method that I
will be able to use to tell me if the line is too long.