Improve wrapping of destructured import statements
As lines get longer, we are currently wrapping destructured import
statements like:

  import { foo, bar, baz, fizz, buzz } from
    'a_really_long_path_for_some_strange_reason';

But I think it makes more sense and is more conventional to wrap
destructured statements like:

  import {
    foo,
    bar,
    baz,
    fizz,
    buzz,
  } from 'a_really_long_path_for_some_strange_reason';

This commit implements this change.

Although this currently works, I am a little unhappy with how tangly
this stuff is in ImportStatement so I plan on following up with some
refactoring that should make it a bit more sane.

Fixes #106.