Fix parsing of new wrapping style of destructuring
I noticed some odd behavior with the new wrapping style of destructuring
imports. Basically, it was not removing unused references from
destructured imports, it was wrapping them weirdly on subsequent runs,
and if you removed some variables from a destructured import and ran fix
on the file again, it would put them back but in a second set of
curlies.

This was all wrong because the regex used to parse these imports was not
in multiline mode. I added some specs that would have prevented this.

Fixes #110 and fixes #111.