Fail better when eslint parser module cannot be read
I was trying out import-js in a project that has an .eslintrc.js file
that relies on a parser installed by npm. When running auto import in
this project, it was simply just not doing anything, which was
perplexing because I could see eslint reporting undefined variables that
import-js could resolve. After some digging, I discovered that this was
because eslint was outputting errors to stdout like:

> <text>:0:0: Cannot find module 'babel-eslint'

This commit doesn't resolve the problem I was encountering, but it at
least makes it easier to understand what is going on.

After knowing this, I was able to resolve my problem by installing the
missing eslint parser package globally.