Add eslint_executable configuration option
If there is a local .eslintrc file that includes dependencies on
eslint-* packages, such as configurations, plugins, or parsers, running
the globally installed eslint can fail if these packages are not also
globally installed. To make this smoother for teams, I am adding a
configuration option that allows people to configure the eslint
executable, which enables them to use the locally installed eslint,
which should be able to find the other dependencies for their project.

Fixes #115.